type as.

+++ sggm/regression_model_helper.py
@@ -16,7 +16,7 @@ def generate_noise_for_model_test(x: torch.Tensor) -> torch.Tensor:
     data_std = x.std(dim=0)
     data_mean = x.mean(dim=0)
 
-    noise = torch.rand(x.shape) - 0.5
+    noise = torch.rand(x.shape).type_as(x) - 0.5
     noise *= (hypercube_max - hypercube_min) * 2 * data_std
     noise += data_mean
